\name{Ombre-package}
\alias{Ombre-package}
\alias{Ombre}
\docType{package}
\title{
Ombre for Knitting and Embroidery
}
\description{
This package allows you to use ggplot to make knitting and embroidery patterns.
}
\details{
\tabular{ll}{
Package: \tab Ombre\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-09\cr
License: \tab What license is it under?\cr
Depends: \tab ggplot2 \cr
}

}
\author{
Margaret Lange

Maintainer: Margaret Lange <margaretmeek@gmail.com>
}

\references{
~~ Literature or other references for background information ~~
}

\keyword{ crafts }

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
~~ simple examples of the most important functions ~~
}
